+++ src/imio/events/core/browser/bring_event_into_agendas/agendas.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 
+from imio.smartweb.common.utils import get_vocabulary
 from imio.smartweb.locales import SmartwebMessageFactory as _
+from plone import api
 from plone.autoform import directives
 from plone.autoform.form import AutoExtensibleForm
 from plone.app.z3cform.widget import AjaxSelectFieldWidget
@@ -35,23 +37,49 @@ class BringEventIntoAgendasForm(AutoExtensibleForm, form.Form):
     schema = IBringEventIntoAgendasForm
     ignoreContext = True
     enable_autofocus = False
-    label = _("Add this event in your agendas")
+    label = _("Add/Remove agenda(s)")
 
-    @buttonAndHandler("Submit")
+    def updateWidgets(self):
+        agendas_to_display = []
+        vocabulary = get_vocabulary("imio.events.vocabulary.UserAgendas")
+        # Loop to display only agenda where user has the permission (ex : to remove these agendas out of event)
+        for agenda_uid in self.context.selected_agendas:
+            if vocabulary.by_token.get(agenda_uid) is None:
+                # user can't remove this agenda because he has no permission on it so we don't display it
+                pass
+            agendas_to_display.append(agenda_uid)
+        self.fields["agendas"].field.default = agendas_to_display
+        super(BringEventIntoAgendasForm, self).updateWidgets()
+
+    @buttonAndHandler(_("Submit"))
     def handle_submit(self, action):
         data, errors = self.extractData()
         if errors:
             self.status = self.formErrorsMessage
             return
-        for new_agenda in data.get("agendas"):
-            if new_agenda not in self.context.selected_agendas:
-                self.context.selected_agendas.append(new_agenda)
-        transaction.commit()
-        self.context.reindexObject(idxs=["selected_agendas"])
-        success_message = _("Event correctly added in agenda(s).")
-        self.request.response.redirect(self.context.absolute_url())
-        self.status = success_message
 
-    @button.buttonAndHandler("Cancel")
+        if len(data.get("agendas")) < len(self.fields["agendas"].field.default):
+            # we want to remove agenda(s) out of this event
+            agendas_to_remove = list(
+                set(self.fields["agendas"].field.default) - set(data.get("agendas"))
+            )
+            for agenda in agendas_to_remove:
+                self.context.selected_agendas.remove(agenda)
+            success_message = _("Agenda(s) correctly removed.")
+        else:
+            # we want to add an agenda in this event
+            for agenda in data.get("agendas"):
+                if agenda not in self.context.selected_agendas:
+                    self.context.selected_agendas.append(agenda)
+            success_message = _("Agenda(s) correctly added.")
+
+        self.context.reindexObject(idxs=["selected_agendas"])
+        transaction.commit()
+        self.status = success_message
+        api.portal.show_message(_(self.status), self.request)
+
+        self.request.response.redirect(self.context.absolute_url())
+
+    @button.buttonAndHandler(_("Cancel"))
     def handleCancel(self, action):
         self.request.response.redirect(self.context.absolute_url())