+++ src/imio/events/core/vocabularies.py
@@ -144,10 +144,6 @@ EventTypesVocabulary = EventTypesVocabularyFactory()
 class UserAgendasVocabularyFactory:
 
     def __call__(self, context=None):
-        current_event = context
-        if not IEvent.providedBy(context) and context is not None:
-            # context is the form
-            current_event = context.context
         site = api.portal.get()
         user = site.portal_membership.getAuthenticatedMember()
         results = []
@@ -156,10 +152,7 @@ class UserAgendasVocabularyFactory:
         for brain in brains:
             obj = brain.getObject()
             try:
-                # Display only agendas where user has the permission to add an event and where event is not already in
-                if user.has_permission(permission, obj) and brain.UID not in getattr(
-                    current_event, "selected_agendas", []
-                ):
+                if user.has_permission(permission, obj):
                     results.append(
                         SimpleTerm(
                             value=brain.UID, token=brain.UID, title=brain.breadcrumb